Modify RefSymbol from IFL0/IFL1 to IFC1/IFC2 Modify RefData field Decimal point from double to int

+++ common/src/main/java/com/cyanspring/common/staticdata/RefData.java
@@ -124,8 +124,8 @@ public class RefData extends DataObject {
 		return this.get(double.class, RefDataField.PRICE_LIMIT.value());
 	}
 	
-	public double getDeciamlPoint(){
-		return this.get(double.class, RefDataField.DECIMALPOINT.value());
+	public int getDeciamlPoint(){
+		return this.get(int.class, RefDataField.DECIMALPOINT.value());
 	}
 	
 	public double getDenominator(){
